+++ pandas/core/window/rolling.py
@@ -103,8 +103,8 @@ if TYPE_CHECKING:
         DataFrame,
         Series,
     )
-    from pandas.core.generic import NDFrame
     from pandas.core.groupby.ops import BaseGrouper
+    from pandas.core.ndframe import NDFrame
 
 
 class BaseWindow(SelectionMixin):
