+++ pandas/tests/ndframe/test_duplicate_labels.py
