+++ pandas/tests/ndframe/test_finalize.py
