+++ scripts/tests/test_no_bool_in_ndframe.py
@@ -1,4 +1,4 @@
-from scripts.no_bool_in_generic import check_for_bool_in_generic
+from scripts.no_bool_in_ndframe import check_for_bool_in_ndframe
 
 BAD_FILE = "def foo(a: bool) -> bool:\n    return bool(0)"
 GOOD_FILE = "def foo(a: bool_t) -> bool_t:\n    return bool(0)"
@@ -6,7 +6,7 @@ GOOD_FILE = "def foo(a: bool_t) -> bool_t:\n    return bool(0)"
 
 def test_bad_file_with_replace():
     content = BAD_FILE
-    mutated, result = check_for_bool_in_generic(content)
+    mutated, result = check_for_bool_in_ndframe(content)
     expected = GOOD_FILE
     assert result == expected
     assert mutated
@@ -14,7 +14,7 @@ def test_bad_file_with_replace():
 
 def test_good_file_with_replace():
     content = GOOD_FILE
-    mutated, result = check_for_bool_in_generic(content)
+    mutated, result = check_for_bool_in_ndframe(content)
     expected = content
     assert result == expected
     assert not mutated
